+++ packages/bridge-status-controller/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `batchId` to BridgeHistoryItem to enable querying history by batchId when txId is not defined ([#6058](https://github.com/MetaMask/core/pull/6058))
+- Update txHistory item with status, approvalTxId and txMeta.id when transactionConfirmed or transactionFailed events are published by the TransactionController ([#6058](https://github.com/MetaMask/core/pull/6058))
 
 ### Changed